test: use `@octokit/auth` for authentication strategies other than token authentication, deprecate previews auth options
