+++ test/issues/1497-include-error-message-on-validation-error-test.js
@@ -8,8 +8,8 @@ describe("https://github.com/octokit/rest.js/issues/1497", () => {
     nock("https://request-errors-test.com", {
       reqheaders: {
         accept:
-          "application/vnd.github.machine-man-preview+json,application/vnd.github.hellcat-preview+json,application/vnd.github.luke-cage-preview+json,application/vnd.github.zzzax-preview+json",
-        authorization: "Bearer dXNlcm5hbWU6cGFzc3dvcmQ="
+          "application/vnd.github.hellcat-preview+json,application/vnd.github.luke-cage-preview+json,application/vnd.github.zzzax-preview+json",
+        authorization: "basic dXNlcm5hbWU6cGFzc3dvcmQ="
       }
     })
       .put("/repos/gr2m/sandbox/branches/gr2m-patch-1/protection", {
@@ -34,7 +34,7 @@ describe("https://github.com/octokit/rest.js/issues/1497", () => {
 
     const octokit = new Octokit({
       baseUrl: "https://request-errors-test.com",
-      auth: "Bearer dXNlcm5hbWU6cGFzc3dvcmQ="
+      auth: "basic dXNlcm5hbWU6cGFzc3dvcmQ="
     });
     return octokit.repos
       .updateBranchProtection({
