+++ test/issues/861-custom-accept-header-test.js
@@ -7,9 +7,8 @@ describe("https://github.com/octokit/rest.js/issues/861", () => {
   it("custom accept header", () => {
     nock("https://issues-861-test.com", {
       reqheaders: {
-        accept:
-          "application/vnd.github.antiope-preview+json,application/vnd.github.machine-man-preview+json",
-        authorization: "Bearer 123"
+        accept: "application/vnd.github.antiope-preview+json",
+        authorization: "token 123"
       }
     })
       .post("/repos/swinton/example/check-runs")
@@ -18,7 +17,7 @@ describe("https://github.com/octokit/rest.js/issues/861", () => {
 
     const octokit = new Octokit({
       baseUrl: "https://issues-861-test.com",
-      auth: "Bearer 123"
+      auth: "token 123"
     });
 
     return octokit.checks.create({
